Moving DistReference before archiving
Previously, if I have an artifact repository that has zip files, with name "jay-logger-1.2.zip", and if I use satis on top of it, it would create an archive file based on the logic in composer to be "jay-logger-1.2-jay-logger-1.2.zip.tar".

The reason being, that the distReference information created by Composer has it as the basename of the file. This causes the file name to be appended between <vendor>-<package>-<distReference>.tar format. If However, the distReference is empty, it would look just fine.