Allow Satis to keep a number of files that would otherwise be pruned
When working with an environment with many changes, a large set of
repositories and/or slow connections from clients that call
packages.json Satis could generate new files for packages whilst clients
were still downloading files from an older version of packages.json.
This caused HTTP 404 Not Found issues because old JSON files were already
pruned. This is occurred in environments with `providers` on `true`.

This change adds a new configuration `providers-history-size` which will
keep a number equal to the setting of old provider-specific JSON's to
remain stored.