Archive file name does not conform to the specified archive format
When the dist type of the remote archive is different with the dist type
specified in satis.json, the remote dist type is used as part of the
file name of the archive generated by satis, take the following scenario
as an example:

```bash
$ cd ~
$ composer create-project composer/satis:dev-master --keep-vcs
$ mkdir satis.server
$ vi satis.server/satis.json
$ # Update satis.json with the following content:
```

```json
{
    "name": "satis/test",
    "homepage": "https://github.com/composer/satis",
    "repositories": [
        { "type": "composer", "url": "https://packagist.org" }
    ],
    "require": {
        "monolog/monolog": "1.13.0"
    },
    "config": {
        "secure-http": false
    },
    "require-dependencies": false,
    "require-dev-dependencies": false,
    "archive": {
        "directory": "dist",
        "format": "tar",
        "prefix-url": "http://satis.localhost:4680",
        "skip-dev": false
    }
}
```

```bash
$ satis/bin/satis satis.server/satis.json satis.server
$ ls satis.server/dist/monolog/monolog
$ # The archive file is as follows:
$ monolog-monolog-c41c218e239b50446fd883acb1ecfd4b770caeae-zip-d4a976.tar
```

> NOTE: The dist type zip of the remote archive is preserved as part of
the archive file name
monolog-monolog-c41c218e239b50446fd883acb1ecfd4b770caeae-zip-d4a976.tar,
this causes a problem when the satis.server is used as an upstream cache
for other satis instances because when a 2nd level ratis server fetches
archive from the upstream one, the archive will be renamed to
monolog-monolog-c41c218e239b50446fd883acb1ecfd4b770caeae-tar-d4a976.tar,
which causes a cache miss every time.

To fix this issue, an option `override-dist-type` has been added to the
`archive` section, when it is set to true, the local dist type takes
precendence over the remote one.

| Question           | Answer
| ------------------ | ------------------
| **issue**          | 478
| **Bug**            | yes
| **New feature**    | no
| **Target version** | `1.x`
| **BC breaks**      | no
| **Tests pass**     | yes
| **Doc needed**     | no