[Composer v2] Add support for available-packages / available-package-patterns
Improves performance as Composer v2 won't need to do 404 requests for packages you don't have (in p2/ folder).
If you have a lot of packages you can configure available-package-patterns to at least narrow the amount of 404's Composer would do.

Currently no way to disable this, unless available-package-patterns is set to omething like ['*']`.

Todo: This needs config for poeple hosting a lot of packages.
Simplest would be if you can provide a pattern or disable completly as alternative to this which can work as default.